[FIX] Catches an exception at shutdown
